requirements.txt
vosk
sounddevice
pywin32
numpy
google-generativeai
python-dotenv
